Move ingredients scroll to a separate hook
